# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Administração de Códigos Promocionais",
        "description": "API para gerenciamento de códigos promocionais.",
        "termsOfService": "http://seusite.com/terms/",
        "contact": {
            "email": "suporte@seusite.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/codigos": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Criação de um novo código promocional",
                "description": "Permite criar um novo código promocional com as informações fornecidas.",
                "operationId": "81a5c21c2228b1dbac9a8470365fc180",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pontuacao",
                                    "resgate_unico"
                                ],
                                "properties": {
                                    "pontuacao": {
                                        "description": "Pontuação associada ao código promocional",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "resgate_unico": {
                                        "description": "Se o código pode ser resgatado uma única vez",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "limite_resgate": {
                                        "description": "Número máximo de resgates permitidos",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "decrescimo": {
                                        "description": "Se o valor do resgate diminui após cada uso",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código promocional criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código promocional criado com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar código promocional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao criar código promocional."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/codigos/{codigoPromocional}": {
            "put": {
                "tags": [
                    "admin"
                ],
                "summary": "Atualização de código promocional",
                "description": "Permite atualizar as informações de um código promocional existente.",
                "operationId": "e067787d1f141b706bb52888d938d387",
                "parameters": [
                    {
                        "name": "codigoPromocional",
                        "in": "path",
                        "description": "ID do código promocional a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pontuacao",
                                    "resgate_unico"
                                ],
                                "properties": {
                                    "pontuacao": {
                                        "description": "Pontuação associada ao código promocional",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "resgate_unico": {
                                        "description": "Se o código pode ser resgatado uma única vez",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "limite_resgate": {
                                        "description": "Número máximo de resgates permitidos",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "decrescimo": {
                                        "description": "Se o valor do resgate diminui após cada uso",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código promocional atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código promocional atualizado com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar código promocional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao atualizar código promocional."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin"
                ],
                "summary": "Exclusão de código promocional",
                "description": "Realiza a exclusão (soft delete) de um código promocional.",
                "operationId": "9134196be9841e0422d2456485ca8ab2",
                "parameters": [
                    {
                        "name": "codigoPromocional",
                        "in": "path",
                        "description": "ID do código promocional a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Código promocional excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código promocional excluído com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao excluir código promocional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao excluir código promocional."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/codigos/{codigoPromocional}/resgatavel": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Verificar se o código é resgatável",
                "description": "Verifica se um código promocional pode ser resgatado.",
                "operationId": "2713c2748b69947aab97986b7fb8547d",
                "parameters": [
                    {
                        "name": "codigoPromocional",
                        "in": "path",
                        "description": "ID do código promocional a ser verificado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O código é resgatável",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O código é resgatável."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "O código não é mais resgatável",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O código não é mais resgatável."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/codigos/{codigoPromocional}/resgatar": {
            "post": {
                "tags": [
                    "competidor"
                ],
                "summary": "Resgatar um código promocional",
                "description": "Permite que o competidor resgate um código promocional.",
                "operationId": "f1f82af7ea83ce86f82eae325c23a24c",
                "parameters": [
                    {
                        "name": "codigoPromocional",
                        "in": "path",
                        "description": "Código promocional a ser resgatado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Código resgatado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código resgatado com sucesso."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao resgatar código promocional",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao resgatar código promocional."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a token gerada no login para acessar rotas protegidas.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}